Add files via upload
9-18-2025
The DrawFunctions DrawObjectPoints2(), DrawObjectPoints2Top(),DrawObjectPoints2Bottom(), DrawObjectPointsTopBottom2() have been added and greatly reduce the number of points and faces as there are no longer duplicates. The ancilliary functions such as CalculateMeshFaces(), CalculateMeshFacesTopBottom(), EndpointLeftFaces(), etc have been modified. DirectXHelper.h ReadAsync() for *.CSO's  is changed and modernized

*.Xaml backgrounds are changed to DarkGreen for readability in Dark Mode and Light Mode Windows

The current DirectXToolkit.xaml is compatible with this version

I have added a ton of Error Checking and will contine to do so.